requirements.txt
ultralytics
torchvision
langchain
langgraph
chromadb
streamlit
python-dotenv
pyyaml
opencv-python
langchain-groq
langchain-community
langchain-huggingface
sentence-transformers
pypdf
langchain-text-splitters
beautifulsoup4
requests
